Problems related the pages have been fixed.

+++ Models/Project.cs
@@ -2,7 +2,7 @@
 {
     public class Project
     {
-        public Int16 Id { get; set; }
+        public int Id { get; set; }
         public String Title { get; set; }
         public Decimal Budget { get; set; }
     }
